fix type in indexes array

+++ algorithm design manual/chapter 3/tasks/use of tree structures/11 - smallest element of subsequence/smallest_friendly_array.cpp
@@ -3,12 +3,12 @@
 void SmallestFriendlyArray::expand_smallest_elements_in_subseq(int new_element){
     // Индекс нового элемента равен размеру elements до вставки
     size_t indx_of_new_element = elements.size();
-    for(std::vector<int> & subseq_for_some_element : smallest_elements_in_subseq){
+    for(std::vector<size_t> & subseq_for_some_element : smallest_elements_in_subseq){
         // Индекс наименьшего элемента в подпоследовательности  x_i ... x_n 
         // равен либо индексу текущему минимума, либо индексу нового элемента
         size_t indx_of_smallest = indx_of_new_element;
         if(!subseq_for_some_element.empty()){
-            int back = subseq_for_some_element.back();
+            size_t back = subseq_for_some_element.back();
             if(elements[back] < new_element) indx_of_smallest = back;
         }
         subseq_for_some_element.push_back(indx_of_smallest);
@@ -22,7 +22,7 @@ void SmallestFriendlyArray::expand_smallest_elements_in_subseq(int new_element){
 SmallestFriendlyArray::SmallestFriendlyArray(const std::vector<int> & vec){
     // Инициализация
     elements = std::vector<int>();
-    smallest_elements_in_subseq = std::vector<std::vector<int>>();
+    smallest_elements_in_subseq = std::vector<std::vector<size_t>>();
 
     for(int element : vec){
         insert(element);
