+++ algorithm design manual/chapter 3/tasks/use of tree structures/11 - smallest element of subsequence/smallest_friendly_array.h
@@ -6,7 +6,7 @@
 class SmallestFriendlyArray{
 private:
     std::vector<int> elements;
-    std::vector<std::vector<int>> smallest_elements_in_subseq;
+    std::vector<std::vector<size_t>> smallest_elements_in_subseq;
     void expand_smallest_elements_in_subseq(int new_element);
 public:
     SmallestFriendlyArray(const std::vector<int> & vec);
